+++ Services/Interfaces/IProgramCourses.cs
@@ -9,9 +9,9 @@ namespace EduReg.Services.Interfaces
         Task<GeneralResponse> CreateProgramCourseAsync(ProgramCoursesDto model);
         Task<GeneralResponse> CreateProgramCourseAsync(List<ProgramCoursesDto> model);
         //Task<GeneralResponse> CreateProgramCourseAsync(byte[] model); // EXCEL, CSV, Flat file upload
-        Task<GeneralResponse> UpdateProgramCourseAsync(int Id, ProgramCoursesDto model);
-        Task<GeneralResponse> DeleteProgramCourseAsync(int Id);
-        Task<GeneralResponse> GetProgramCoursesByIdAsync(int Id);
+        Task<GeneralResponse> UpdateProgramCourseAsync(long Id, ProgramCoursesDto model);
+        Task<GeneralResponse> DeleteProgramCourseAsync(long Id);
+        Task<GeneralResponse> GetProgramCoursesByIdAsync(long Id);
         Task<GeneralResponse> GetProgramCoursesByProgramNameAsync(string programName);//Shortname
         Task<GeneralResponse> GetAllProgramsByCoursesAsync();
     }
