+++ Services/Repositories/FeeItemsRepository.cs
@@ -88,7 +88,7 @@ namespace EduReg.Services.Repositories
 
 
         // ✅ Get Fee Item by Id
-        public async Task<GeneralResponse> GetFeeItemByIdAsync(int Id)
+        public async Task<GeneralResponse> GetFeeItemByIdAsync(long Id)
         {
             var item = await _context.FeeItem
                 .Include(x => x.FeeRules)
@@ -113,7 +113,7 @@ namespace EduReg.Services.Repositories
 
 
         // ✅ Update Fee Item
-        public async Task<GeneralResponse> UpdateFeeItemAsync(int Id, FeeItemDto model)
+        public async Task<GeneralResponse> UpdateFeeItemAsync(long Id, FeeItemDto model)
         {
             var existingItem = await _context.FeeItem.FindAsync(Id);
 
@@ -159,7 +159,7 @@ namespace EduReg.Services.Repositories
 
 
         // ✅ Delete Fee Item
-        public async Task<GeneralResponse> DeleteFeeItemAsync(int Id)
+        public async Task<GeneralResponse> DeleteFeeItemAsync(long Id)
         {
             var item = await _context.FeeItem.FindAsync(Id);
