+++ Services/Repositories/FeeRulesRepository.cs
@@ -63,7 +63,7 @@ namespace EduReg.Services.Repositories
             return new GeneralResponse { StatusCode = 200, Message = "Fee rules retrieved successfully.", Data = rules };
         }
 
-        public async Task<GeneralResponse> GetFeeRuleByIdAsync(int id, string institutionShortName)
+        public async Task<GeneralResponse> GetFeeRuleByIdAsync( long id, string institutionShortName)
         {
             var rule = await _context.FeeRule
                 .Include(r => r.FeeItem)
@@ -81,7 +81,7 @@ namespace EduReg.Services.Repositories
 
         }
 
-        public async Task<GeneralResponse> UpdateFeeRuleAsync(int id, FeeRuleDto model, string institutionShortName)
+        public async Task<GeneralResponse> UpdateFeeRuleAsync(long id, FeeRuleDto model, string institutionShortName)
         {
             var rule = await _context.FeeRule
                 .FirstOrDefaultAsync(r => r.Id == id && r.InstitutionShortName == institutionShortName);
@@ -107,7 +107,7 @@ namespace EduReg.Services.Repositories
 
         }
 
-        public async Task<GeneralResponse> DeleteFeeRuleAsync(int id, string institutionShortName)
+        public async Task<GeneralResponse> DeleteFeeRuleAsync(long id, string institutionShortName)
         {
             var rule = await _context.FeeRule
                 .FirstOrDefaultAsync(r => r.Id == id && r.InstitutionShortName == institutionShortName);
