+++ Services/Repositories/RegistrationsBusinessRulesRepository.cs
@@ -279,7 +279,7 @@ namespace EduReg.Services.Repositories
             }
         }
 
-        public async Task<GeneralResponse> UpdateRegistrationBusinessRuleAsync(int Id, RegistrationBusinessRulesDto model)
+        public async Task<GeneralResponse> UpdateRegistrationBusinessRuleAsync(long Id, RegistrationBusinessRulesDto model)
         {
             try
             {
@@ -329,7 +329,7 @@ namespace EduReg.Services.Repositories
             }
         }
 
-        public async Task<GeneralResponse> DeleteRegistrationBusinessRuleAsync(int Id)
+        public async Task<GeneralResponse> DeleteRegistrationBusinessRuleAsync(long Id)
         {
             try
             {
